Safari-611.1.10/LayoutTests/accessibility/aria-listbox-clear-selection-crash-expected.txt

Option 1
Option 2
Tests that attempting to clear the selection in an ARIA listbox doesn't cause a crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS listbox.selectedChildrenCount is 0
PASS successfullyParsed is true

TEST COMPLETE
